+++ src/plone/restapi/tests/test_types.py
@@ -63,6 +63,18 @@ class ITaggedValuesSchema(model.Schema):
         description=u"",
     )
 
+    form.read_permission(field_read_protected='cmf.ManagePortal')
+    field_read_protected = schema.TextLine(
+        title=u"ReadProtected",
+        description=u"",
+    )
+
+    form.write_permission(field_write_protected='cmf.ManagePortal')
+    field_write_protected = schema.TextLine(
+        title=u"WriteProtected",
+        description=u"",
+    )
+
 
 class TestJsonSchemaUtils(TestCase):
 
@@ -187,6 +199,24 @@ class TestTaggedValuesJsonSchemaUtils(TestCase):
         #     False,
         #     jsonschema['properties']['field_omitted_false']['omitted']
         # )
+
+    def test_get_jsonschema_with_field_permissions(self):
+        ttool = getToolByName(self.portal, 'portal_types')
+        jsonschema = get_jsonschema_for_fti(
+            ttool['TaggedDocument'],
+            self.portal,
+            self.request
+        )
+
+        self.assertEqual(
+            'cmf.ManagePortal',
+            jsonschema['properties']['field_read_protected']['read_permission']
+        )
+
+        self.assertEqual(
+            'cmf.ManagePortal',
+            jsonschema['properties']['field_write_protected']['write_permission']  # noqa
+        )
 
 
 class TestJsonSchemaProviders(TestCase):